+++ src/number/is-network-port/index.ts
@@ -1,0 +1,2 @@
+export { IsNetworkPort, IS_NETWORK_PORT } from './is-network-port.decorator'
+export { isNetworkPort } from './is-network-port.predicate'